refactor(core): コア層のサブモジュール化（bootstrap, clipboard, events）の完遂とそれに伴うインポートパスの修正・整理
