+++ reusable_gui/components/history_list_component.py
@@ -5,6 +5,7 @@ from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
     from src.core.base_application import BaseApplication
+
     from src.gui.base.context_menu import HistoryContextMenu
