+++ reusable_gui/components/phrase_list_component.py
@@ -6,6 +6,7 @@ from tkinter import messagebox
 from typing import TYPE_CHECKING, cast
 
 from src.core.exceptions import PhraseError
+
 from src.gui.base import context_menu
 from src.gui.base.base_frame_gui import BaseFrameGUI
 
@@ -13,6 +14,7 @@ logger = logging.getLogger(__name__)
 
 if TYPE_CHECKING:
     from src.core.base_application import BaseApplication
+
     from src.gui.base.context_menu import PhraseListContextMenu
     from src.gui.components.phrase_edit_component import PhraseEditComponent
