+++ src/core/bootstrap/base_application.py
