+++ src/core/bootstrap/dependency_checker.py
