+++ src/core/bootstrap/exceptions.py
