+++ src/core/clipboard/__init__.py
@@ -1,0 +1,1 @@
+# clipboard initialization
