+++ src/core/events/__init__.py
@@ -1,0 +1,1 @@
+# events initialization
