+++ src/gui/base/base_frame_gui.py
@@ -6,7 +6,7 @@ from typing import TYPE_CHECKING, Any
 from src.utils.error_handler import log_and_show_error
 
 if TYPE_CHECKING:
-    from src.core.base_application import BaseApplication
+    from src.core.bootstrap.base_application import BaseApplication
 
 
 class BaseFrameGUI(tk.Frame):
