+++ src/gui/components/history_list_component.py
@@ -4,7 +4,7 @@ import tkinter as tk
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from src.core.base_application import BaseApplication
+    from src.core.bootstrap.base_application import BaseApplication
     from src.gui.base.context_menu import HistoryContextMenu
