+++ src/gui/dialogs/phrase_edit_dialog.py
@@ -8,7 +8,7 @@ from src.gui.base.base_toplevel_gui import BaseToplevelGUI
 from src.gui.custom_widgets import CustomText
 
 if TYPE_CHECKING:
-    from src.core.base_application import BaseApplication
+    from src.core.bootstrap.base_application import BaseApplication
     from src.services.fixed_phrases_manager import FixedPhrasesManager
