+++ src/gui/windows/meta_management_window.py
@@ -15,7 +15,7 @@ from src.utils.error_handler import log_and_show_error
 
 if TYPE_CHECKING:
     from src.core.app_main import MainApplication
-    from src.core.base_application import BaseApplication
+    from src.core.bootstrap.base_application import BaseApplication
 
 logger = logging.getLogger(__name__)
