+++ src/plugins/implementations/hash_calculator_plugin.py
@@ -11,7 +11,7 @@ from src.gui.base.base_frame_gui import BaseFrameGUI
 from src.plugins.base_plugin import Plugin
 
 if TYPE_CHECKING:
-    from src.core.base_application import BaseApplication
+    from src.core.bootstrap.base_application import BaseApplication
 
 
 # The GUI Component implementation, moved from gui/components
