+++ src/plugins/implementations/unit_converter_plugin.py
@@ -10,7 +10,7 @@ from src.gui.base.base_frame_gui import BaseFrameGUI
 from src.plugins.base_plugin import Plugin
 
 if TYPE_CHECKING:
-    from src.core.base_application import BaseApplication
+    from src.core.bootstrap.base_application import BaseApplication
 
 
 class UnitConverterComponent(BaseFrameGUI):
