+++ tests/conftest.py
@@ -4,7 +4,7 @@ from collections.abc import Generator
 
 import pytest
 
-from src.core.event_dispatcher import EventDispatcher
+from src.core.events.event_dispatcher import EventDispatcher
 from src.db.database_manager import DatabaseManager
 from src.services.history_service import HistoryService
